0.0.1a8: opt-out of sqlite in main db
